+++ SpatialConnect/SCRemoteDataStore.m
@@ -24,13 +24,13 @@
 }
 
 - (void)listenForConnection {
-  //TODO
-//  SCSensorService *ss = [[SpatialConnect sharedInstance] sensorService];
-//  [[[ss.isConnected filter:^BOOL(NSNumber *x) {
-//    return x.boolValue;
-//  }] take:1] subscribeNext:^(NSNumber *x) {
-//    self.status = SC_DATASTORE_RUNNING;
-//  }];
+  // TODO
+  //  SCSensorService *ss = [[SpatialConnect sharedInstance] sensorService];
+  //  [[[ss.isConnected filter:^BOOL(NSNumber *x) {
+  //    return x.boolValue;
+  //  }] take:1] subscribeNext:^(NSNumber *x) {
+  //    self.status = SC_DATASTORE_RUNNING;
+  //  }];
 }
 
 - (void)resume {
