+++ SpatialConnect/SCService.h
@@ -20,14 +20,6 @@
 #import "SCServiceLifecycle.h"
 #import "SCServiceStatusEvent.h"
 
-typedef NS_ENUM(NSInteger, SCServiceStatus) {
-  SC_SERVICE_STARTED,
-  SC_SERVICE_PAUSED,
-  SC_SERVICE_RUNNING,
-  SC_SERVICE_STOPPED,
-  SC_SERVICE_ERROR
-};
-
 @interface SCService : NSObject <SCServiceLifecycle>
 
 @property(readonly, nonatomic) SCServiceStatus status;
