+++ SpatialConnect/SCServiceLifecycle.h
@@ -20,6 +20,14 @@
 #import <Foundation/Foundation.h>
 #import <ReactiveCocoa/ReactiveCocoa.h>
 
+typedef NS_ENUM(NSInteger, SCServiceStatus) {
+  SC_SERVICE_STARTED,
+  SC_SERVICE_PAUSED,
+  SC_SERVICE_RUNNING,
+  SC_SERVICE_STOPPED,
+  SC_SERVICE_ERROR
+};
+
 @protocol SCServiceLifecycle <NSObject>
 
 @required
@@ -29,5 +37,6 @@
 - (void)stop;
 - (void)startError:(NSError *)e;
 - (NSArray *)requires;
+- (SCServiceStatus)status;
 
 @end
