+++ SpatialConnect/SCServiceNode.h
@@ -10,18 +10,21 @@
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and limitations under the License
+ * See the License for the specific language governing permissions and
+ * limitations under the License
  */
 
-#import <Foundation/Foundation.h>
 #import "SCService.h"
-#import "SCServiceEdge.h"
+#import <Foundation/Foundation.h>
+
+@class SCServiceEdge;
 
 @interface SCServiceNode : NSObject
 
-- (id)initWithService:(id<SCServiceLifecycle>)s andEdges:(NSArray<SCServiceEdge*>*)e;
+@property(nonatomic, readonly) id<SCServiceLifecycle> service;
+@property(nonatomic, readonly) NSArray<SCServiceEdge *> *edges;
 
-@property (nonatomic,readonly) id<SCServiceLifecycle> service;
-@property (nonatomic,readonly) NSArray<SCServiceEdge*>* edges;
+- (id)initWithService:(id<SCServiceLifecycle>)s
+             andEdges:(NSArray<SCServiceEdge *> *)e;
 
 @end
